+++ src/Entity/League.php
@@ -13,6 +13,8 @@ use Drupal\Core\Entity\EntityTypeInterface;
 use Drupal\mespronos\MPNEntityInterface;
 use Drupal\Core\Database\Database;
 use Drupal\file\Entity\File;
+use Drupal\Core\Url;
+use Drupal\mespronos\Entity\RankingGeneral;
 
 /**
  * Defines the League entity.
@@ -30,6 +32,7 @@ use Drupal\file\Entity\File;
  *       "default" = "Drupal\mespronos\Entity\Form\LeagueForm",
  *       "add" = "Drupal\mespronos\Entity\Form\LeagueForm",
  *       "edit" = "Drupal\mespronos\Entity\Form\LeagueForm",
+ *       "archive" = "Drupal\mespronos\Entity\Form\LeagueArchiveForm",
  *       "delete" = "Drupal\mespronos\Entity\Form\MPNDeleteForm",
  *     },
  *     "access" = "Drupal\mespronos\ControlHandler\LeagueAccessControlHandler",
@@ -45,6 +48,7 @@ use Drupal\file\Entity\File;
  *   links = {
  *     "edit-form" = "/entity.league.edit_form",
  *     "recount_points" = "/entity.league.recount_points",
+ *     "archive" = "/entity.league.archive",
  *     "delete-form" = "/entity.league.delete_form",
  *     "collection" = "/entity.league.collection"
  *   },
@@ -110,7 +114,14 @@ class League extends MPNContentEntityBase implements MPNEntityInterface {
     return count($ids);
   }
 
-
+  public function getBettersNumber() {
+    $injected_database = Database::getConnection();
+    $query = $injected_database->select('mespronos__ranking_league','rl');
+    $query->addExpression('count(rl.better)','nb_better');
+    $query->condition('rl.league',$this->id());
+    $results = $query->execute()->fetchObject();
+    return $results->nb_better;
+  }
   /**
    * Return all days for league
    * @return \Drupal\mespronos\Entity\Day[]
@@ -228,6 +239,18 @@ class League extends MPNContentEntityBase implements MPNEntityInterface {
     }
   }
 
+  public function getRenderableLabel() {
+
+    return [
+      '#theme' => 'league',
+      '#league' => [
+        'url' => Url::fromRoute('mespronos.league.index',['league'=>$this->id()]),
+        'label' => $this->label(),
+        'logo' => $this->getLogo('mini_logo')
+      ]
+    ];
+  }
+
   public function getLogo($style_name = 'thumbnail') {
     $logo = $this->get("field_league_logo")->first();
     if($logo && !is_null($logo) && $logo_file = File::load($logo->getValue()['target_id'])) {
@@ -237,7 +260,6 @@ class League extends MPNContentEntityBase implements MPNEntityInterface {
       return [];
     }
   }
-
 
   public function getPoints() {
     $points = [
@@ -250,6 +272,13 @@ class League extends MPNContentEntityBase implements MPNEntityInterface {
 
   public function isActive() {
     return $this->get('status')->value == 'active';
+  }
+
+  public function close() {
+    $this->set('status','archived');
+    $this->save();
+    \Drupal::logger('mespronos')->notice(t('League @league_label as been set as archived',['@league_label'=>$this->label()]));
+    RankingGeneral::createRanking();
   }
 
   /**